backend/requirements.txt
anyio==3.7.0
bcrypt==4.0.1
cachetools==5.3.1
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.95.2
filelock==3.7.1
google-auth==2.20.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
isort==5.10.1
jaraco.context==4.2.0
lazy-object-proxy==1.7.1
mccabe==0.7.0
more-itertools==9.0.0
mysql-connector-python==8.0.31
mysqlclient==2.1.1
passlib==1.7.4
Pillow==9.2.0
pipenv==2022.5.2
platformdirs==2.5.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.8
pylint==2.13.9
python-decouple==3.6
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
sqlparse==0.4.2
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.6.2
tzdata==2022.1
urllib3==1.26.14
uvicorn==0.22.0
virtualenv==20.14.1
virtualenv-clone==0.5.7
watchfiles==0.19.0
websockets==11.0.3
wrapt==1.14.1
